+++ packages/deepsea-tools/CHANGELOG.md
@@ -1,5 +1,11 @@
 # deepsea-tools
 
+## 5.5.0
+
+### Minor Changes
+
+- tree 的方法变更
+
 ## 5.4.0
 
 ### Minor Changes